Tidy up the server file
